refactor(vtc): move role-change orchestration into ceremony/ (P2.1, part 1)
The role-change `decide → effect` spine (`role_change_via_pipeline` +
`assemble_role_change_facts` + `RoleChangeResult`) lived in
`routes/members/update.rs` but is shared by two entry points (the PATCH handler
and the promote-to-admin step-up). Move it — and its P0.14 policy tests — into a
new `ceremony::orchestrate` module so the orchestration sits beside the pipeline
it drives, is unit-testable without axum, and no longer needs a
`crate::routes::…` back-reference from `promote.rs`.

`update_member` and `promote_finish` are now thin adapters: they call
`crate::ceremony::role_change_via_pipeline` and keep only their own
(divergent) post-effect audit — `RoleChanged` vs `AdminPromoted` — so the audit
correctly stays at the call site while the policy-load / decide / effect / facts
logic leaves the route layer.

Behaviour-preserving: the moved P0.14 tests pass in their new home; members_crud
(13) + full lib (582) green; fmt + clippy clean. This establishes the
`ceremony/orchestrate.rs` home; the leave (`remove_inner`) and join
(`submit_inner`) spines follow.
